+++ pages/TrustandSecurity.py
@@ -14,3 +14,4 @@ def show():
     st.write("""
     Note: If you have further questions about Streamlit Trust and Security, please reach out to them on the [Streamlit Community Forum](https://discuss.streamlit.io/). 💬
     """)
+
