+++ requirements.txt
@@ -1,87 +1,233 @@
+aiofiles==22.1.0
 aiohttp==3.8.5
 aiosignal==1.3.1
+aiosqlite==0.19.0
 altair==5.0.1
 altgraph==0.17.3
+analytics-python==1.4.post1
+anyio==3.7.1
+appdirs==1.4.4
+appnope==0.1.3
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
+arrow==1.2.3
+asttokens==2.2.1
 async-timeout==4.0.2
 attrs==23.1.0
+Babel==2.12.1
+backcall==0.2.0
+backoff==1.10.0
+beautifulsoup4==4.12.2
+bleach==6.0.0
 blinker==1.6.2
+blis==0.7.10
 cachetools==5.3.1
+catalogue==2.0.9
 certifi==2023.7.22
+cffi==1.15.1
+chardet==3.0.4
 charset-normalizer==3.2.0
 click==8.1.6
+colorama==0.4.6
+comm==0.1.4
+confection==0.1.1
 contourpy==1.1.0
 cycler==0.11.0
+cymem==2.0.7
+debugpy==1.6.7.post1
 decorator==5.1.1
+defusedxml==0.7.1
+en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
+entrypoints==0.4
 et-xmlfile==1.1.0
+executing==1.2.0
+Faker==19.3.0
+fastjsonschema==2.18.0
+favicon==0.7.0
 filelock==3.12.2
 fonttools==4.40.0
+forex-python==1.8
+fqdn==1.5.1
+frozendict==2.3.8
 frozenlist==1.4.0
 fsspec==2023.6.0
+gensim==4.3.1
 gitdb==4.0.10
 GitPython==3.1.32
+googletrans==4.0.0rc1
+h11==0.9.0
+h2==3.2.0
+hpack==3.0.0
+hstspreload==2023.1.1
+htbuilder==0.6.1
+html5lib==1.1
+httpcore==0.9.1
+httpx==0.13.3
 huggingface-hub==0.16.4
-idna==3.4
+hyperframe==5.2.0
+idna==2.10
 importlib-metadata==6.8.0
+ipykernel==6.25.1
+ipython==8.14.0
+ipython-genutils==0.2.0
+isoduration==20.11.0
+jedi==0.19.0
 Jinja2==3.1.2
 joblib==1.3.1
+json5==0.9.14
+jsonpointer==2.4
 jsonschema==4.18.4
 jsonschema-specifications==2023.7.1
+jupyter-events==0.7.0
+jupyter-ydoc==0.2.5
+jupyter_client==7.4.9
+jupyter_core==5.3.1
+jupyter_server==2.7.0
+jupyter_server_fileid==0.9.0
+jupyter_server_terminals==0.4.4
+jupyter_server_ydoc==0.8.0
+jupyterlab==3.6.5
+jupyterlab-pygments==0.2.2
+jupyterlab_server==2.24.0
 kiwisolver==1.4.4
+langcodes==3.3.0
+langdetect==1.0.9
+lxml==4.9.3
 macholib==1.16.2
+Markdown==3.4.4
 markdown-it-py==3.0.0
+markdownlit==0.0.7
 MarkupSafe==2.1.3
 matplotlib==3.7.2
+matplotlib-inline==0.1.6
 mdurl==0.1.2
+mistune==3.0.1
+mitoinstaller==0.0.237
+mitosheet==0.1.503
+monotonic==1.6
+more-itertools==10.1.0
 multidict==6.0.4
+multitasking==0.0.11
+murmurhash==1.0.9
+nbclassic==1.0.0
+nbclient==0.8.0
+nbconvert==7.7.3
+nbformat==5.9.2
+nest-asyncio==1.5.7
+nltk==3.8.1
+notebook==6.5.5
+notebook_shim==0.2.3
+NRCLex==3.0.0
 numpy==1.25.1
+oauthlib==3.2.2
 openai==0.27.8
 openpyxl==3.1.2
+overrides==7.4.0
 packaging==23.1
 pandas==2.0.3
+pandocfilters==1.5.0
+parso==0.8.3
+pathy==0.10.2
 pdfkit==1.0.0
+pexpect==4.8.0
+pickleshare==0.7.5
 Pillow==9.5.0
+platformdirs==3.10.0
 plotly==5.15.0
+preshed==3.0.8
+prometheus-client==0.17.1
+prompt-toolkit==3.0.39
 protobuf==4.23.4
+psutil==5.9.5
+ptyprocess==0.7.0
+pure-eval==0.2.2
 pyarrow==12.0.1
+pycparser==2.21
+pydantic==1.10.12
 pydeck==0.8.0
 Pygments==2.15.1
 pyinstaller==5.13.0
 pyinstaller-hooks-contrib==2023.5
+pymdown-extensions==10.1
 Pympler==1.0.1
 pyparsing==3.0.9
 python-dateutil==2.8.2
+python-json-logger==2.0.7
 pytz==2023.3
 pytz-deprecation-shim==0.1.0.post0
 PyYAML==6.0.1
+pyzmq==24.0.1
 referencing==0.30.0
 regex==2023.6.3
 requests==2.31.0
+requests-oauthlib==1.3.1
+rfc3339-validator==0.1.4
+rfc3986==1.5.0
+rfc3986-validator==0.1.1
 rich==13.4.2
 rpds-py==0.9.2
 safetensors==0.3.1
 scikit-learn==1.3.0
 scipy==1.11.1
 seaborn==0.12.2
+Send2Trash==1.8.2
+simplejson==3.19.1
 six==1.16.0
+smart-open==6.3.0
 smmap==5.0.0
+sniffio==1.3.0
+soupsieve==2.4.1
+spacy==3.6.0
+spacy-legacy==3.0.12
+spacy-loggers==1.0.4
+srsly==2.4.7
+st-annotated-text==4.0.0
+stack-data==0.6.2
 streamlit==1.25.0
+streamlit-camera-input-live==0.2.0
+streamlit-card==0.0.61
+streamlit-embedcode==0.1.2
+streamlit-extras==0.3.0
+streamlit-faker==0.0.2
+streamlit-image-coordinates==0.1.6
+streamlit-keyup==0.2.0
 streamlit-option-menu==0.3.6
+streamlit-toggle-switch==1.0.2
+streamlit-vertical-slider==1.0.2
 tenacity==8.2.2
+termcolor==2.3.0
+terminado==0.17.1
+textblob==0.17.1
+thinc==8.1.10
 threadpoolctl==3.2.0
+tinycss2==1.2.1
 tokenizers==0.13.3
 toml==0.10.2
 toolz==0.12.0
 tornado==6.3.2
 tqdm==4.65.0
+traitlets==5.9.0
 transformers==4.31.0
+tweepy==4.14.0
+typer==0.9.0
 typing_extensions==4.7.1
 tzdata==2023.3
 tzlocal==4.3.1
+uri-template==1.3.0
 urllib3==2.0.4
 validators==0.20.0
+wasabi==1.1.2
 watchdog==3.0.0
+wcwidth==0.2.6
+webcolors==1.13
+webencodings==0.5.1
+websocket-client==1.6.1
 wkhtmltopdf==0.2
+wordcloud==1.9.2
 xlrd==2.0.1
-XlsxWriter==3.1.2
+XlsxWriter==3.0.2
+y-py==0.6.0
 yarl==1.9.2
+yfinance==0.2.27
+ypy-websocket==0.8.4
 zipp==3.16.2
